stop returning builders in built-in setter api
The built-in setters will adopt a new pattern that allows the user
to pass in an error handler, otherwise a default is used. This means
it can just return a regular setter now. It sucked to call the setter
method on every flag anyways.

The original idea here was to return a version of the setter with type
information intact, so the user can override the function used to
convert the error to an plain english message. They can do that either
way of course, but this way they get compile errors if the wrong error
set is handled.

This small thing is nice to have because sourceopt wants to avoid
forcing any particular wording on the end user, since they should have
complete control over the visual output of the program.